requirements.txt
# FastAPI and web framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# AWS SDK
boto3==1.35.0
botocore==1.35.0

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.36

# OpenSearch
opensearch-py==2.7.1
requests-aws4auth==1.3.1

# Configuration and validation
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1

# HTTP client for image downloads
httpx==0.27.2

# Image processing
Pillow==11.0.0

# Utilities
tenacity==9.0.0

# Logging
structlog==24.4.0

# Development (optional)
pytest==8.3.3
pytest-asyncio==0.24.0
